I have implemented concurrent query support and a multi-threaded chat UI for QCX. Here are the key updates:
- I transitioned the query server action to be non-blocking, allowing it to return immediately while processing AI responses in the background.
- I introduced a `ChatThreadManager` to manage multiple independent chat sessions and tabs.
- I isolated the `AIState` and persistence by partitioning them with a `threadId`.
- I improved streaming reliability by initializing `StreamableValue` with empty strings and ensuring that all streams are properly concluded.
- I integrated the `CalendarNotepad` as a toggleable overlay within the `ChatView`.
- I refactored common chat logic, such as `getChatTitle`, into utility functions.
